# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * OSINT Intelligence Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface ScanRequest {
  /** Domain or IP address to scan */
  target: string;
  /** Optional list of modules to run (dns, whois, shodan, subdomains, emails, blacklist, technologies) */
  modules?: string[];
}

export interface DnsRecord {
  type: string;
  value: string;
  ttl?: number;
}

export interface WhoisData {
  registrar?: string;
  registrantOrg?: string;
  registrantCountry?: string;
  createdDate?: string;
  expiresDate?: string;
  updatedDate?: string;
  nameServers?: string[];
  status?: string[];
  emails?: string[];
}

export interface ShodanService {
  port: number;
  transport?: string;
  product?: string;
  version?: string;
  banner?: string;
  cpe?: string[];
}

export interface ShodanHost {
  ip?: string;
  org?: string;
  isp?: string;
  country?: string;
  city?: string;
  os?: string;
  ports?: number[];
  vulns?: string[];
  tags?: string[];
  hostnames?: string[];
  services?: ShodanService[];
  asn?: string;
  lastUpdate?: string;
}

export interface EmailResult {
  email: string;
  firstName?: string;
  lastName?: string;
  position?: string;
  department?: string;
  confidence?: number;
  sources?: string[];
  linkedin?: string;
}

export interface SubdomainResult {
  subdomain: string;
  ip?: string;
  status?: number;
  title?: string;
}

export interface BlacklistResult {
  listed: boolean;
  listCount?: number;
  lists?: string[];
  details?: string;
}

export interface TechnologyResult {
  name: string;
  category?: string;
  version?: string;
  confidence?: number;
  website?: string;
}

export interface IpNetblockResult {
  ip?: string;
  netblock?: string;
  asn?: string;
  asnName?: string;
  country?: string;
  org?: string;
}

export type ScanResultScanType =
  (typeof ScanResultScanType)[keyof typeof ScanResultScanType];

export const ScanResultScanType = {
  domain: "domain",
  ip: "ip",
} as const;

export type ScanResultEmailStats = {
  total?: number;
  pattern?: string;
  organization?: string;
};

export type ScanResultErrors = { [key: string]: string };

export interface ScanResult {
  target: string;
  scanType: ScanResultScanType;
  timestamp: string;
  dns?: DnsRecord[];
  whois?: WhoisData;
  shodan?: ShodanHost[];
  subdomains?: SubdomainResult[];
  emails?: EmailResult[];
  emailStats?: ScanResultEmailStats;
  blacklist?: BlacklistResult;
  technologies?: TechnologyResult[];
  ipNetblocks?: IpNetblockResult[];
  errors?: ScanResultErrors;
}

export interface SavedScan {
  id: number;
  target: string;
  scanType: string;
  timestamp: string;
  result: ScanResult;
}