# lib/api-zod/src/generated/types/blacklistResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * OSINT Intelligence Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface BlacklistResult {
  listed: boolean;
  listCount?: number;
  lists?: string[];
  details?: string;
}