# lib/api-zod/src/generated/types/scanResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * OSINT Intelligence Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { BlacklistResult } from "./blacklistResult";
import type { DnsRecord } from "./dnsRecord";
import type { EmailResult } from "./emailResult";
import type { IpNetblockResult } from "./ipNetblockResult";
import type { ScanResultEmailStats } from "./scanResultEmailStats";
import type { ScanResultErrors } from "./scanResultErrors";
import type { ScanResultScanType } from "./scanResultScanType";
import type { ShodanHost } from "./shodanHost";
import type { SubdomainResult } from "./subdomainResult";
import type { TechnologyResult } from "./technologyResult";
import type { WhoisData } from "./whoisData";

export interface ScanResult {
  target: string;
  scanType: ScanResultScanType;
  timestamp: string;
  dns?: DnsRecord[];
  whois?: WhoisData;
  shodan?: ShodanHost[];
  subdomains?: SubdomainResult[];
  emails?: EmailResult[];
  emailStats?: ScanResultEmailStats;
  blacklist?: BlacklistResult;
  technologies?: TechnologyResult[];
  ipNetblocks?: IpNetblockResult[];
  errors?: ScanResultErrors;
}