# lib/api-zod/src/generated/types/shodanHost.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * OSINT Intelligence Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { ShodanService } from "./shodanService";

export interface ShodanHost {
  ip?: string;
  org?: string;
  isp?: string;
  country?: string;
  city?: string;
  os?: string;
  ports?: number[];
  vulns?: string[];
  tags?: string[];
  hostnames?: string[];
  services?: ShodanService[];
  asn?: string;
  lastUpdate?: string;
}