# lib/api-zod/src/generated/types/technologyResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * OSINT Intelligence Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface TechnologyResult {
  name: string;
  category?: string;
  version?: string;
  confidence?: number;
  website?: string;
}